requirements-dev.txt
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-default-group==1.2.2
cryptography==36.0.2
decorator==5.1.1
docopt==0.6.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
imgkit==1.2.2
outcome==1.1.0
pipreqs==0.4.11
playwright==1.21.0
pycodestyle==2.8.0
pycparser==2.21
pyee==8.1.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.26.0
selenium==4.1.3
shot-scraper==0.13
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urlbox==1.0.6
urllib3==1.26.9
validators==0.18.2
webdriver-manager==3.5.4
websockets==10.1
wsproto==1.1.0
yarg==0.1.9
